requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-crispy-forms==1.14.0
django-import-export==3.0.2
django-jazzmin==2.6.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.0
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0